Move Glyph tooltip story
